+++ src/Enums/AppStatus.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Busarm\PhpMini\Enums;
+
+/**
+ * PHP Mini Framework
+ *
+ * @copyright busarm.com
+ * @license https://github.com/Busarm/php-mini/blob/master/LICENSE (MIT License)
+ * @codeCoverageIgnore
+ */
+class AppStatus
+{
+    const INITIALIZING  =   -1;
+    const RUNNNIG       =   1;
+    const STOPPED     =   0;
+}