+++ tests/TestApp/Services/MockStatelessService.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Busarm\PhpMini\Test\TestApp\Services;
+
+use Busarm\PhpMini\Interfaces\SingletonStatelessInterface;
+use Busarm\PhpMini\Traits\SingletonStateless;
+
+class MockStatelessService implements SingletonStatelessInterface
+{
+    use SingletonStateless;
+    public function __construct(public $id)
+    {
+    }
+}